ARROW-9910: [Rust][DataFusion] Fixed error in type coercion of Variadic.
The background of this error is that we currently use `numerical_coercion(lhs, rhs)` function to check if a coercion rule applies, but that function tries to come up with a common type between lhs and rhs. We should be using the function `can_coerce_from` to check if a cast can be performed.

This commit fixes this.

Closes #8108 from jorgecarleitao/type_fix

Authored-by: Jorge C. Leitao <jorgecarleitao@gmail.com>
Signed-off-by: Paddy Horan <paddyhoran@hotmail.com>