ARROW-9753: [Rust] [DataFusion] Replaced Arc<Mutex<>> by Box<>
This PR is built on top of #8225 and Replaces `Arc<Mutex<dyn ...>>` by `Box<dyn ...>`.

In the TopK example, I had to move some functions away from the `impl`. This is because `self` cannot be borrowed as mutable and immutable at the same time, and, during iteration, it was being borrowed as mutable (to update the BTree) and as immutable (to access the `input`). There is probably a better way of achieving this e.g. via interior mutability.

Closes #8307 from jorgecarleitao/box_iterator

Authored-by: Jorge C. Leitao <jorgecarleitao@gmail.com>
Signed-off-by: Andy Grove <andygrove73@gmail.com>